codegen: leave out `.toShort` to avoid accidental integer overflow (#329)
if the numbers get too large we want the compiler to tell us,
really... compare:

```scala
def foo(s: Short) = print(s)

foo(5)
// 5

foo(500000)
-- [E007] Type Mismatch Error: --------
  |foo(500000)
  |    ^^^^^^
  |    Found:    (500000 : Int)
  |    Required: Short
```

+++ test-schemas-domain-classes/src/main/scala/testdomains/codepropertygraphminified/nodes/Call.scala
@@ -43,7 +43,7 @@ object Call {
 }
 
 class Call(graph_4762: flatgraph.Graph, seq_4762: Int)
-    extends StoredNode(graph_4762, 0.toShort, seq_4762)
+    extends StoredNode(graph_4762, 0, seq_4762)
     with CallBase
     with CallRepr
     with StaticType[CallEMT] {
@@ -142,7 +142,7 @@ object NewCall {
   }
 }
 
-class NewCall extends NewNode(0.toShort) with CallBase with CallReprNew {
+class NewCall extends NewNode(0) with CallBase with CallReprNew {
   override type StoredNodeType = Call
   override def label: String = "CALL"
