+++ test-schemas-domain-classes/src/main/scala/testdomains/generic/nodes/NodeA.scala
@@ -63,7 +63,7 @@ object NodeA {
 }
 
 class NodeA(graph_4762: flatgraph.Graph, seq_4762: Int)
-    extends StoredNode(graph_4762, 0.toShort, seq_4762)
+    extends StoredNode(graph_4762, 0, seq_4762)
     with NodeABase
     with StaticType[NodeAEMT] {
   def node_b: Option[NodeB] = flatgraph.Accessors.getNodePropertyOption[NodeB](graph, nodeKind, 6, seq)
@@ -272,7 +272,7 @@ object NewNodeA {
   }
 }
 
-class NewNodeA extends NewNode(0.toShort) with NodeABase {
+class NewNodeA extends NewNode(0) with NodeABase {
   override type StoredNodeType = NodeA
   override def label: String = "node_a"
