+++ test-schemas-domain-classes/src/main/scala/testdomains/gratefuldead/edges/EdgeTypes.scala
@@ -6,7 +6,7 @@ object Followedby {
 }
 
 class Followedby(src_4762: flatgraph.GNode, dst_4762: flatgraph.GNode, subSeq_4862: Int, property_4862: Any)
-    extends flatgraph.Edge(src_4762, dst_4762, 0.toShort, subSeq_4862, property_4862) {
+    extends flatgraph.Edge(src_4762, dst_4762, 0, subSeq_4862, property_4862) {
   override def propertyName: Option[String] = Followedby.propertyName
 }
 
@@ -16,7 +16,7 @@ object Sungby {
 }
 
 class Sungby(src_4762: flatgraph.GNode, dst_4762: flatgraph.GNode, subSeq_4862: Int, property_4862: Any)
-    extends flatgraph.Edge(src_4762, dst_4762, 1.toShort, subSeq_4862, property_4862) {}
+    extends flatgraph.Edge(src_4762, dst_4762, 1, subSeq_4862, property_4862) {}
 
 object Writtenby {
   val Label = "writtenBy"
@@ -24,4 +24,4 @@ object Writtenby {
 }
 
 class Writtenby(src_4762: flatgraph.GNode, dst_4762: flatgraph.GNode, subSeq_4862: Int, property_4862: Any)
-    extends flatgraph.Edge(src_4762, dst_4762, 2.toShort, subSeq_4862, property_4862) {}
+    extends flatgraph.Edge(src_4762, dst_4762, 2, subSeq_4862, property_4862) {}
